# Fix: Trump Tractor Performance Issue with Smart Combination Generator